Add modules for game cell and field logic
